Fix WindowStack to correctly handle removing windows of the same type
For now, the window stack in the js layer is virtual and not represented
on the backend. There is only a single window on the backend to
represent each type, so a window with a window of the same type beneath
being popped in js should trigger a content change rather than a pop in
the backend.