Move action enabled bool back to window action props
Toggling the state of the action bar causes the window's click config to
be set to either an action bar click config or scroll layer click
config, both of which are not desirable for a menu which requires the
menu click config.

For now, ensure that all window operations the menu requires do not in
any way relate to the action bar, thus this property must be grouped
with actions. As a result, windows are responsible for clearing the
action bar themselves in their respective clearing procedures.