Add aplite and SDK 2.9 compatibility header
Rather than undefining the compatibility headers provided by SDK 3.0 and
redefining them, copy all missing definitions and fallback on those
defined in the header only if they are missing.

The header is also an aplite compatibility header, giving aplite a few
definitions that are only available on basalt such as GColor8 and
GBitmapFormat.