Unconditionally render the background color as if on basalt
To the furthest extent possible, ifdefs should only be in header files
in order to simplify the codebase. Aplite has access to the full
background GColor, use it anyway. Eventually there should be a color
filter based on readability that would keep the background white on
aplite unless the text color is black.