Add GColor8GetOr for maintaining color readability on aplite
Many color combinations are readable on basalt, but aplite only supports
black and white. To support cross compiling pebble.js apps, fallback to
a white background in the case that the user specifies a different
background color, otherwise the background will be clear. Only the card
required this as it uses a black background to implement rounded
corners. The menu already has a white background.