Invert palette black and white images in menu. Addresses #55
The SDK is converting black and white PNGs, and the firmware is loading
these PNGs as 1-bit PNGs with a palette rather than without. This is
causing the firmware drawing logic to draw them as color PNGs when they
are originally purely black and white. Menu Layers also no longer invert
their selection.

To support the usage of 1-bit images, simulate a faux inversion.