Fix button coloring for 1-bit images in the action bar. Related to #55
The action bar does not invert images that have palettes. PNGs without
palettes are being loaded with palettes.

Change the palette of black and white images to have faux inversion when
used on the action bar. This has the caveat that instances of an image
that is simultaneously used both inside and outside an action bar will
render with unexpected colors that only make sense in the action bar.