Fix scheduling a wakeup event immediately. Addresses #82
The Wakeup module was created when the timezone API was exposed, but not
yet enabled. This required falling back to the phone for calculating the
corrected timestamp if communicating with a non-timezone Pebble.
However, scheduling a wakeup doesn't wait until the fallback logic
decides whether it is necessary in the first place, causing issues with
immediately scheduling a startup with timezone enabled Pebbles.

The fix wraps the schedule message in a Wakeup launch callback,
deferring the schedule process until the Pebble has communicated whether
it supports timezones or not.