Fix window stack handling on basalt
Basalt animates windows by animating the window directly, therefore
performing multiple window popping and pushing simultaneously can cause
memory mismanagement edge cases because of unexpected usage of the
window stack. This fix disables the pusher on basalt. The pusher is the
additional window that is used in aplite to manipulate the window stack
to perform the desired window animation.