Fix menu to send props only after the window is pushed
Menu was sending props twice, once before the menu is loaded and once
after. This is because the menu properties packet is responsible for
both background color and the number of sections. Menu has multiple
layers of resolving in order to dynamically update menus, but since
the number of sections is in the same packet as background color,
resolving the menu before the window is shown has the unintended side
affect of sending menu window properties twice.

This fix changes the highest layer of menu resolving to be aware that it
can also push the window so that the window property handling can use
the menu resolving as well. In the future, menu props may be separated,
and window and menu properties may be additionally cached in simply
window and menu.