Load JavaScript vendor scripts as CommonJS modules
moment has always been bundled with pebble.js for convenience, but using
it directly without require is an odd combination with the module style of
pebble.js. Previously, vendor scripts were executed directly in order to
require no modification for inclusion, but this encourages bad
practices. Change the build to wrap vendor scripts as CommonJS modules.