Work around Pebble.sendAppMessage double callback call issue
If the JavaScript thread immediately moves onto a job that takes longer
than a few seconds after successfully sending an app message, upon
returning from the long job, Pebble.sendAppMessage will call the failure
callback with a timed out error even though it has already called the
success handler before the long job. MessageQueue expects
Pebble.sendAppMessage calls and callback calls to be one-to-one and will
fail when it is skewed either way.

This fixes the issue by giving the callbacks references to the message
they were sent with. If the last sent message of the queue itself is
different from the reference given to the callback, then it is an extra
callback call and is subsequently ignored.