Don't reset the scroll layer if it is not being used
SimplyMenu does not use the scroll layer, therefore resetting the scroll
layer will break SimplyMenu scrolling due to the scroll layer taking
over the click config. Add a SimplyWindow boolean property that
explicitly specifies whether the scroll layer is used in order to avoid
resetting and applying the scroll layer click config if it is not used.