Fix structs string length members to use a StringLengthType transform (Addresses #159)
structs were using EnumerableType which was overloaded to handle
strings. However, when a number type is used in place of where a string
is expected, StringType and EnumerableType don't cooperate. StringType
will convert the number correctly, but EnumerableType will just return
the number value itself.

The fix is to move the string length logic out of EnumerableType into a
new type StringLengthType. StringType was also modified to treat
undefined as the empty string, since undefined would then show up for
unspecified text fields.